transactional_app/requirements.txt
kafka-python==2.0.1
google-cloud-storage==2.14.0
